requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
dj-database-url==2.0.0
Django==4.1.7
django-appconf==1.0.5
django-crispy-forms==2.0
django-filter==23.1
django-watermark==0.2.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2==2.9.5
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.7
whitenoise==6.5.0
